# backend/internal/api/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SwiftInbox External Developer API",
    "version": "1.0.0",
...
  },
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/webhooks": {
      "get": {
        "tags": ["Webhooks"],
        "summary": "List webhooks",
        "description": "Lists all webhooks owned by the authenticated developer.",
        "operationId": "listWebhooks",
        "x-usage": { "operation": "webhook.get", "credits": 1, "category": "api" },
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "The developer's webhooks.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Webhook" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      },
      "post": {
        "tags": ["Webhooks"],
        "summary": "Create a webhook",
        "description": "Creates a webhook and links it to one or more owned mailboxes. The returned webhook includes the one-time `secret` used to verify signatures in the `X-Webhook-Signature` header.",
        "operationId": "createWebhook",
        "x-usage": { "operation": "webhook.create", "credits": 5, "category": "api" },
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/WebhookCreateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook created.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Webhook" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
...
    "/api/v1/webhooks/{id}": {
      "get": {
        "tags": ["Webhooks"],
        "summary": "Get a webhook",
        "description": "Fetches a single webhook owned by the authenticated developer by its UUID.",
        "operationId": "getWebhook",
        "x-usage": { "operation": "webhook.get", "credits": 1, "category": "api" },
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Webhook UUID.",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "The webhook.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Webhook" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      },
      "patch": {
        "tags": ["Webhooks"],
        "summary": "Update a webhook",
        "description": "Updates the `name` and/or `is_active` fields of a webhook. Omitted fields keep their current values.",
        "operationId": "updateWebhook",
        "x-usage": { "operation": "webhook.update", "credits": 1, "category": "api" },
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Webhook UUID.",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/WebhookUpdateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated webhook.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Webhook" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      },
      "delete": {
        "tags": ["Webhooks"],
        "summary": "Delete a webhook",
        "description": "Deletes a webhook owned by the authenticated developer.",
        "operationId": "deleteWebhook",
        "x-usage": { "operation": "webhook.delete", "credits": 1, "category": "api" },
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Webhook UUID.",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Webhook deleted.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/WebhookDeleted" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
...
    "/api/v1/webhooks/{id}/test": {
      "post": {
        "tags": ["Webhooks"],
        "summary": "Send a test event",
        "description": "Delivers a test `webhook.test` event to the webhook endpoint and reports the delivery result. Testing an inactive webhook returns `409 Conflict`.",
        "operationId": "testWebhook",
        "x-usage": { "operation": "webhook.test", "credits": 2, "category": "api" },
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Webhook UUID.",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Delivery attempt finished.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/WebhookTestResult" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "dev_<64 hex chars>",
        "description": "API key sent as a Bearer token in the `Authorization` header (the middleware also accepts the legacy `Authentication` header):\n\n```\nAuthorization: Bearer dev_<64 hex characters>\n```\n\nCreate or manage keys in the developer dashboard (**Api keys** page)."
      }
    },
    "responses": {
      "BadRequest": {
        "description": "The request is missing or malformed (invalid path value, invalid JSON body, wrong domain, invalid expiry, etc.).",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing, malformed, unknown, or revoked API key.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Forbidden": {
        "description": "The authenticated developer is not allowed to access the requested resource (for example another developer's mailbox).",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "NotFound": {
        "description": "The requested mailbox, message, attachment, webhook, or dead letter does not exist (or does not belong to the developer).",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Conflict": {
        "description": "The request conflicts with the current state of the resource (mailbox already exists, mailbox already linked, webhook inactive).",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "TooManyRequests": {
        "description": "The developer's usage quota was exceeded (`{\"error\":\"usage quota exceeded\"}`) or the IP rate limit was reached.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "InternalServerError": {
        "description": "An unexpected server error occurred.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message."
          }
        },
        "required": ["error"]
      },
...
      "WebhookCreateRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "description": "Optional label; defaults to the endpoint URL when omitted." },
          "url": { "type": "string", "format": "uri", "example": "https://example.com/webhooks/receiver" },
          "events": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/WebhookEvent" },
            "description": "Events the webhook subscribes to. At least one is required."
          },
          "mailbox_ids": {
            "type": "array",
            "items": { "type": "string", "format": "uuid" },
            "description": "Mailboxes to link to the webhook. At least one is required."
          }
        },
        "required": ["url", "events", "mailbox_ids"]
      },
      "WebhookUpdateRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "is_active": { "type": "boolean", "description": "Pause or resume delivery of events." }
        }
      },
...
      "Webhook": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "developer_id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "url": { "type": "string", "format": "uri" },
          "is_active": { "type": "boolean" },
          "events": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/WebhookEvent" }
          },
          "secret": {
            "type": "string",
            "description": "Webhook signing secret sent in the `X-Webhook-Signature` header of deliveries. Only returned when a webhook is first created."
          },
          "mailbox_ids": {
            "type": "array",
            "items": { "type": "string", "format": "uuid" }
          }
        },
        "required": ["id", "developer_id", "name", "url", "is_active", "events", "mailbox_ids"]
      },
      "WebhookEvent": {
        "type": "string",
        "enum": ["email.received", "email.deleted", "mailbox.expired"],
        "description": "The events a webhook can subscribe to."
      },
      "WebhookDeleted": {
        "type": "object",
        "properties": {
          "message": { "type": "string", "example": "webhook deleted" },
          "id": { "type": "string", "format": "uuid" }
        },
        "required": ["message", "id"]
      },
...
      "WebhookTestResult": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "True when the endpoint replied with a 2xx status."
          },
          "statusCode": { "type": "integer", "description": "HTTP status code returned by the webhook endpoint." }
        },
        "required": ["success", "statusCode"]
      },
...
    }
  }
}